requirements.txt
fastapi
uvicorn[standard]
pandas
openpyxl
xlrd>=2.0.1
python-multipart
sqlalchemy
pydantic-settings
pyodbc
